config/win32/mingw.mak: enable warnings config/win32/config.h: add HAVE_STRFTIME lib/win32/sys.c: quiet DWORD formatting complaints

+++ lib/win32/sys.c
@@ -227,14 +227,14 @@ osname(cp)
 	break;
 
     default:
-	sprintf(osname, "Win#%d", osv.dwPlatformId);
+	sprintf(osname, "Win#%d", (int)osv.dwPlatformId);
 	os = osname;
 	vnum = 1;
 	break;
     } // dwPlatformId
 
     if (vnum) {
-	sprintf(cp, "%s %d.%d", os, osv.dwMajorVersion, osv.dwMinorVersion);
+	sprintf(cp, "%s %d.%d", os, (int)osv.dwMajorVersion, (int)osv.dwMinorVersion);
 	if (build) {
 	    cp += strlen(cp);
 	    sprintf(cp, ".%d", build);
